perf(www): move heavy route deps out of the critical import graph
Route files' module-level bindings (docs clientLoader, demos barrel helper, preview examples cache) ride in the router's critical graph on every page. Moving them to non-route -page files cuts the landing entry from 454 to 215 KB gz and preloads from 140 to 46. Also deep-import the create/preset barrel to keep pako out of the root graph, and request GitHub avatars at 96px instead of full size.
